refactor: move Safe Context Menu components from old Sidebar folder to My Accounts (the consumer)
