+++ apps/web/src/features/myAccounts/components/MyAccountsV2/components/AccountItem/MultiAccountItem.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import NextLink from 'next/link'
 import Identicon from '@/components/common/Identicon'
 import FiatBalance from '@/features/spaces/components/SelectSafesOnboarding/components/FiatBalance'
-import MultiAccountContextMenu from '@/components/sidebar/SafeListContextMenu/MultiAccountContextMenu'
+import MultiAccountContextMenu from '@/features/myAccounts/components/SafeListContextMenu/MultiAccountContextMenu'
 import { AccountItem as BaseAccountItem } from '../../../AccountItem'
 import { AddNetworkButton } from '../../../AddNetworkButton'
 import { useMultiAccountItemData } from '../../../../hooks/useMultiAccountItemData'
