+++ apps/web/src/features/myAccounts/components/SafeListContextMenu/index.tsx
@@ -7,7 +7,7 @@ import MenuItem from '@mui/material/MenuItem'
 import ListItemText from '@mui/material/ListItemText'
 
 import EntryDialog from '@/components/address-book/EntryDialog'
-import SafeListRemoveDialog from '@/components/sidebar/SafeListRemoveDialog'
+import SafeListRemoveDialog from './SafeListRemoveDialog'
 import NestedSafesIcon from '@/public/images/sidebar/nested-safes-icon.svg'
 import EditIcon from '@/public/images/common/edit.svg'
 import DeleteIcon from '@/public/images/common/delete.svg'
@@ -20,7 +20,7 @@ import { AppRoutes } from '@/config/routes'
 import router from 'next/router'
 import { CreateSafeOnNewChain } from '@/features/multichain'
 import { useOwnersGetSafesByOwnerV1Query } from '@safe-global/store/gateway/AUTO_GENERATED/owners'
-import { NestedSafesPopover } from '../NestedSafesPopover'
+import { NestedSafesPopover } from '@/components/sidebar/NestedSafesPopover'
 import { NESTED_SAFE_EVENTS, NESTED_SAFE_LABELS } from '@/services/analytics/events/nested-safes'
 import { useHasFeature } from '@/hooks/useChains'
 import { useNestedSafesVisibility } from '@/hooks/useNestedSafesVisibility'
